Lockheed1049h: re-model the engines completely for better historical accuracy
The modeling of the engines was a bit inaccurate, partly because of
inconsistent data sources and partly due to old limitations of JSBsim.
This new model uses consistent historical data, assuming 135/145 octane
fuel which is no longer available anywhere.  The data used to tune the
model are included with references.