Lockheed1049h version 2.5.2: Implement RMI needles using property rules
The existing implementation used Nasal scripting to overcome an
issue with the in-range flag which was fixed by commit 56912c17
in Oct 2011.

Filtering in a property rule creates a smoother animation of the
RMI needle when switching stations, during station passage and
when the signal is poor.

Patch from Richard Senior.