B-1B: make the navigation display obey the battery switch.
When the battery switch is off, turn the navigation display off.

Note that the real B-1B does not have such a navigation display; it
has (rather: had) an electromechanical horizontal situation indicator
similar to that of the A-10.  After 2001, B-1Bs were upgraded with
partial glass cockpits replacing the monochrome CRT on top, the HSI at
the bottom, and the tape gauges on both sides with two vertical color
LCDs side by side.