MiG-15: disable autopilot, delete unused files
There was a very simple way to break the aircraft in the air and crash
without any hope of recovery: type Ctrl+W, or Ctrl+A, or any of the
default key bindings that enable the autopilot; or, use the F11
autopilot dialog and enable any mode.  The aircraft would go
completely wild, control surfaces would freeze immediately and
disabling the autopilot would not help.

The real MiG-15 did not have an autopilot.  This patch removes the
autopilot files and disables the default key bindings for safety.

Also, delete the unused Panels/ directory and its contents.